fix: add algorithms folder; fix runtime issues; add debug catch block
